requirements-python.txt
# atlasMemory Python Requirements

# TiDB Vector SDK
tidb-vector>=0.0.9

# Database libraries
SQLAlchemy>=2.0.0
pymysql>=1.1.0

# Environment variables
python-dotenv>=1.0.0

# Local embeddings
sentence-transformers>=2.2.0

# Testing
pytest>=7.4.0
